+++ Assets/UniGLTF/Runtime/UniGLTF/IO/TextureIO/ColorSpace.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f3929edbda61f9346906bfab93411b98
+guid: 1a3edb24329fd454db97cac12f30c0d8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
