+++ Assets/UniGLTF/Runtime/UniGLTF/IO/TextureIO/TextureExporter.cs.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: eca1330a83e17a14eb99fc6ea1697922
-timeCreated: 1533533316
-licenseType: Free
+guid: 65fdfff6cc4b1e14a882259e903fc830
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
