+++ Assets/UniGLTF/Runtime/UniGLTF/IO/TextureIO/TextureSamplerUtil.cs.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 435499f173753ac418331fe0f967b789
-timeCreated: 1541561421
-licenseType: Free
+guid: 62f23c5ca623a9f4083c25a63b2c82af
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
