+++ Assets/UniGLTF/Tests/UniGLTF/MaterialTests.cs
@@ -19,7 +19,7 @@ namespace UniGLTF
                 filterMode = FilterMode.Bilinear,
             };
 
-            var textureManager = new TextureExportManager();
+            var textureManager = new TextureExporter();
             var srcMaterial = new Material(Shader.Find("Standard"));
 
             var offset = new Vector2(0.3f, 0.2f);
@@ -255,7 +255,7 @@ namespace UniGLTF
             material.SetColor("_EmissionColor", new Color(0, 1, 2, 1));
             material.EnableKeyword("_EMISSION");
             var materialExporter = new MaterialExporter();
-            var textureExportManager = new TextureExportManager();
+            var textureExportManager = new TextureExporter();
             var gltfMaterial = materialExporter.ExportMaterial(material, textureExportManager);
 
             Assert.AreEqual(gltfMaterial.emissiveFactor, new float[] { 0, 0.5f, 1 });
