+++ Assets/VRM.Samples/Editor/Tests/VRMMaterialTests.cs
@@ -10,7 +10,7 @@ namespace VRM.Samples
         {
             var material = Resources.Load<Material>(resourceName);
             var exporter = new VRMMaterialExporter();
-            var textureManager = new UniGLTF.TextureExportManager();
+            var textureManager = new UniGLTF.TextureExporter();
             var exported = exporter.ExportMaterial(material, textureManager);
 
             // parse glTFExtensionExport to glTFExtensionImport
